reafactors createTrends into separate pages for easier mantainability
