+++ server/config/cron/searchTrends.js
@@ -1,4 +1,9 @@
+const googleTrends = require('google-trends-api');
+// timezone compared to UTC
+const TIMEZONE = -18;
+
 const searchTrends = function(input, date) {
+  console.log('searchTrends');
   const weekAgo = new Date(date - 6 * 24 * 60 * 60 * 1000);
   return googleTrends.interestOverTime({
     keyword: [input],
@@ -8,3 +13,5 @@ const searchTrends = function(input, date) {
     granularTimeResolution: true
   });
 };
+
+module.exports = searchTrends;
